refactor: Move 'getBadges' into DesignerAdapter & DesignerPort
