+++ peauty-customer-api/src/main/java/com/peauty/customer/business/designer/DesignerPort.java
@@ -1,0 +1,10 @@
+package com.peauty.customer.business.designer;
+
+import com.peauty.domain.designer.Badge;
+
+import java.util.List;
+
+public interface DesignerPort {
+
+    List<Badge> getBadges(Long userId);
+}
