+++ peauty-customer-api/src/main/java/com/peauty/customer/business/workspace/WorkspacePort.java
@@ -1,6 +1,5 @@
 package com.peauty.customer.business.workspace;
 
-import com.peauty.domain.designer.Badge;
 import com.peauty.domain.designer.Designer;
 import com.peauty.domain.designer.Rating;
 import com.peauty.domain.designer.Workspace;
@@ -12,6 +11,5 @@ public interface WorkspacePort {
     List<Workspace> findAllWorkspaceByAddress(String address);
     Designer findDesignerById(Long designerId);
     Rating getRatingByWorkspaceId(Long workspaceId); // 추가된 메서드
-    List<Badge> getBadges(Long userId);
 
 }
